@@ -1,33 +1,22 @@
 {
     "private": true,
+    "type": "module",
     "scripts": {
-        "dev": "npm run development",
-        "development": "mix",
-        "watch": "mix watch",
-        "watch-poll": "mix watch -- --watch-options-poll=1000",
-        "hot": "mix watch --hot",
-        "prod": "npm run production",
-        "production": "mix --production"
+        "dev": "vite",
+        "build": "vite build",
+        "prod": "vite build"
     },
     "dependencies": {
-        "axios": "^0.21.0",
-        "moment": "^2.28.0"
+        "axios": "^1.6.7",
+        "moment": "^2.30.1"
     },
     "devDependencies": {
-        "babel-eslint": "^10.1.0",
-        "bootstrap": "^5.1.3",
-        "cross-env": "^7.0.2",
-        "eslint": "^7.10.0",
-        "eslint-loader": "^4.0.2",
-        "eslint-plugin-vue": "^7.0.1",
-        "laravel-mix": "^6.0.43",
-        "laravel-mix-purgecss": "^6.0.0",
-        "resolve-url-loader": "^4.0.0",
-        "sass": "^1.26.11",
-        "sass-loader": "^10.0.2",
-        "turndown": "^7.0.0",
-        "vue": "^2.0.0",
-        "vue-loader": "^15.9.8",
-        "vue-template-compiler": "^2.0.0"
+        "@vitejs/plugin-vue": "^5.0.3",
+        "resolve-url-loader": "^5.0.0",
+        "sass": "~1.70.0",
+        "vite": "^5.1.1",
+        "vue": "^3.4.18",
+        "laravel-vite-plugin": "^1.0.1"
+
     }
 }
